requirements.txt
aiofiles==22.1.0
aiogram==3.0.0b6
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
asyncpg==0.26.0
attrs==22.1.0
Babel==2.9.1
certifi==2022.9.24
charset-normalizer==2.1.1
frozenlist==1.3.1
idna==3.4
magic-filter==1.0.9
multidict==6.0.2
psycopg2-binary==2.9.4
pydantic==1.10.2
python-dotenv==0.21.0
pytz==2022.5
requests==2.28.1
typing_extensions==4.4.0
urllib3==1.26.12
yarl==1.8.1
